Add more tests for Material Offers endpoints

+++ src/routes/material-offer-routes.js
@@ -6,7 +6,7 @@ const MaterialOffer = require('../models/material-offer')
 const MaterialBooking = require('../models/material-booking')
 const Member = require('../models/member')
 
-// S-03b Ritorna i dettagli dell’istanza
+// S-03b
 router.get('/:id', (req, res, next) => {
   if (!req.user_id) {
     return res.status(401).send('Unauthorized')
@@ -17,15 +17,16 @@ router.get('/:id', (req, res, next) => {
       return res.status(404).send("Offer doesn't exist")
     }
 
-    Member.findOne({ user_id: req.user_id, group_id: offer.group_id }).then(() => {
+    Member.findOne({ user_id: req.user_id, group_id: offer.group_id }).then((member) => {
+      if (!member) {
+        return res.status(401).send('Unauthorized')
+      }
       res.json(offer)
-    }).catch((error) => {
-      return res.status(404).send("Offer doesn't exist")
     })
   }).catch(next)
 })
 
-// S-04c Aggiorna l’istanza
+// S-04c
 router.put('/:id', async (req, res, next) => {
   if (!req.user_id) {
     return res.status(401).send('Unauthorized')
@@ -43,35 +44,40 @@ router.put('/:id', async (req, res, next) => {
     }).catch(next)
 })
 
-// S-05b Elimina l’istanza
+// S-05b
 router.delete('/:id', (req, res, next) => {
   if (!req.user_id) {
     return res.status(401).send('Unauthorized')
   }
   const id = req.params.id
-  MaterialOffer.deleteOne( // TODO delete bookings also
-    { material_offer_id: id, created_by: req.user_id })
+  MaterialOffer.deleteOne({ material_offer_id: id, created_by: req.user_id })
     .then(result => {
       if (!result.deletedCount) {
         return res.status(404).send("Offer doesn't exist")
       }
-      res.json(true)
+      MaterialBooking.deleteMany({ offer_id: id }).then(() => {
+        res.json(true)
+      }).catch(next)
     }).catch(next)
 })
 
-// S-08b Associa utente e date di prenotazione
+// S-08b
 router.post('/:id/book', (req, res, next) => {
   if (!req.user_id) {
     return res.status(401).send('Unauthorized')
   }
   const id = req.params.id
   const filter = { material_offer_id: id, created_by: { '$ne': req.user_id } }
+  const bookingReq = req.body
 
   MaterialOffer.findOne(filter).then((offer) => {
+    if (!offer) {
+      return res.status(404).send("Offer doesn't exist")
+    }
     const booking = {
       material_booking_id: objectid(),
-      start: req.start,
-      end: req.end,
+      start: bookingReq.start,
+      end: bookingReq.end,
       user: req.user_id,
       offer_id: offer.material_offer_id
     }
@@ -86,38 +92,51 @@ router.post('/:id/book', (req, res, next) => {
   }).catch(next)
 })
 
-// S-10b Cambia lo stato da “in prestito” a “disponibile” e viceversa
+// S-10b
 router.post('/:id/booked', (req, res, next) => {
   if (!req.user_id) {
     return res.status(401).send('Unauthorized')
   }
   const id = req.params.id
+  const bookingState = req.body
+
   MaterialOffer.findOne({ material_offer_id: id, created_by: { '$eq': req.user_id } })
     .then(offer => {
       if (!offer) {
         return res.status(404).send('Offer not found')
       }
       const filter = { material_offer_id: id }
-      const update = { borrowed: !offer.borrowed }
+      const update = { borrowed: bookingState.borrowed }
 
-      MaterialOffer.updateOne(filter, update)
-      res.json(true)
+      MaterialOffer.updateOne(filter, update).then((offer) => {
+        res.json(offer.borrowed)
+      }).catch(next)
     }).catch(next)
 })
 
-// S-07a Ritorna le prenotazioni del materiale
+// S-07a
 router.get('/:id/bookings', (req, res, next) => {
-  const offer_id = req.query.id
-  MaterialBooking.find({ offer_id: { $in: offer_id } })
-    .lean()
-    .exec()
-    .then(bookings => {
-      if (bookings.length === 0) {
-        return res.status(404).send('No bookings were found')
+  if (!req.user_id) {
+    return res.status(401).send('Unauthorized')
+  }
+
+  const offer_id = req.params.id
+  MaterialOffer.findOne({ material_offer_id: offer_id, created_by: { '$eq': req.user_id } })
+    .then(offer => {
+      if (!offer) {
+        return res.status(404).send('Offer not found')
       }
-      return res.json(bookings)
-    })
-    .catch(next)
+      MaterialBooking.find({ offer_id: offer_id })
+        .lean()
+        .exec()
+        .then(bookings => {
+          if (bookings.length === 0) {
+            return res.status(404).send('No bookings were found')
+          }
+          return res.json(bookings)
+        })
+        .catch(next)
+    }).catch(next)
 })
 
 module.exports = router
